backend/requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==1.10.13
python-multipart==0.0.6
pytest==7.4.3
email-validator
